+++ compiam/__init__.py
@@ -1,4 +1,5 @@
-from compiam.data import models_dict
+import mirdata
+from compiam.data import models_dict, datasets_list
 from compiam.exceptions import ModelNotDefinedError
 
 def load_model(model_name, models_dict=models_dict):
@@ -11,6 +12,25 @@ def load_model(model_name, models_dict=models_dict):
 
     return m_dict['wrapper'](**m_dict['kwargs'])
 
+def load_dataset(dataset_name, data_home=None, version="default"):
+    """Alias function to load a mirdata Dataset class
+    Args:
+        dataset_name (str): the dataset's name
+            see mirdata.DATASETS for a complete list of possibilities
+        data_home (str or None): path where the data lives. If None
+            uses the default location.
+        version (str or None): which version of the dataset to load.
+            If None, the default version is loaded.
+    Returns:
+        Dataset: a mirdata.core.Dataset object
+    """
+    if dataset_name not in datasets_list:
+        raise ValueError("Invalid dataset {}".format(dataset_name)) 
+    return mirdata.initialize(dataset_name=dataset_name, data_home=data_home, version=version)
 
-def load_dataset():
-    pass
+def list_datasets():
+    """Get a list of all mirdata dataset names
+    Returns:
+        list: list of dataset names as strings
+    """
+    return datasets_list
