+++ compiam/data.py
@@ -21,3 +21,8 @@ models_dict = {
 		'kwargs': {}
 	}
 }
+
+###############
+# Datasets List
+###############
+datasets_list = ['saraga_carnatic', 'saraga_hindustani', 'mridangam_stroke']
