--- compiam/datasets/README.md
@@ -1,1 +1,0 @@
-# Datasets
